requirements.txt
BeautifulSoup==3.2.1
Django==1.4
Flask==0.7.2
Jinja2==2.6
Werkzeug==0.7.1
argparse==1.2.1
certifi==0.0.8
chardet==1.0.1
facebook-sdk==0.3.1
mechanize==0.2.5
requests==0.10.4
simplejson==2.2.1
wsgiref==0.1.2
